Improve consistency of the testutil.Set() pattern
While working on the fix for issue #1674 I noticed that most, but not all,
uses of testutil.Set() explicitly modified a pointer to a function being
mocked. However, there are a few cases that use a different pattern.
This changes those cases so future readers of the code aren't wondering
why there are two patterns for using testutil.Set(). As a bonus this allows
removing one source file that exists only to enable the alternative pattern
being eliminated.

This change also includes a few renaming changes that aren't strictly
speaking necessary (e.g., `osExit` => `OSExit`) but are included for
consistency with similar cases.