Implement `path:stat`
I originally started work on this to provide a mechanism to verify the
behavior of other unit tests; e.g., a `path:chmod` command I'm working on.
Primarily because the external `stat` command has different option names
depending on whether you're on a BSD or Linux platform making it hard to
use portably. However, since it seems like this could be generally useful I
went ahead and implemented support for the metadata that is found on some,
not all, platforms.

TBD is providing better support for interpretating the `mode` value.

Related: #1659