Align `pprint` map values
This modifies the behavior of `pprint` to vertically align map values
using the maximum key width rather than using a tab character which often
results in vertically misaligned values. It also changes the indent from
one to two spaces to improve readability of the output.

This also replaces the existing text dedent function with a less rigid
implementation. The new implementation makes it easier to write indented
text in a natural manner that (a) should reliably produce the expected
result, and (b) not catasrophically fail if the input has odd
indentation. In the latter case the expectation is that using the
dedented text will detect it has an unexpected representation. But the
new implementation also produces what I would consider correct output
where the old implementation would panic. Ultimately I think this dedent
functionality should be exposed by the `str` module but didn't want to
go that far with this change.

Related: #1495