Implement a well defined ordering among types
This builds on the previous changes to implement a total ordering by
replacing the dependency on the address at which various types are defined
(which is effectively random) with an explicit ordering. For types without
an explicit ordering it uses an ordering based on the order in which each
unknown type is compared. This makes the ordering predictable based on the
execution of Elvish code without regard to unpredictable factors such as
the layout of data structures in the Elvish binary.

Note that the use of mechanisms like the `peach` command can introduce
unpredictable ordering of unknown value types since this change handles
unknown types at run time and thus, in theory, produce different results
when Elvish code that invokes `compare &total` or `order &total` is run
concurrently.  Whereas the prior logic is not affected by the parallelism of
Elvish code.  Nonetheless, I maintain that the predictability of ordering
heterogenous values introduced by this change is a net positive since the
new behavior will be predictable and reproducible (as a practical matter)
while the prior logic (based on the random layout of data structures in
the Elvish binary) is neither.

Related #1495