Add a useful customizations page
Introduce a new Elvish document that suggests changes to a user's *rc.elv*
likely to be useful by most Elvish users. There are a large number of other
customizations I rely on but most of them are too niche to include in a
document like this one. The intent is to introduce a page that documents
customizations that a significant percentage of Elvish users are likely
to find useful and adopt. Arguably this could be a FAQ page but it seems
to me a page devoted to commonly used customizations is distinct from the
sort of information typically found on a FAQ.