.cirrus.yml: Disable ASLR on FreeBSD so that the tests actually run.
go test -race is not compatible with ASLR, but a current Go bug causes the test
to still appear as successful: https://github.com/golang/go/issues/65425.

This commit should result in a test failure on Cirrus CI's FreeBSD runner,
because there's currently a FreeBSD-specific bug in a test case (#1763).