pkg/buildinfo: Move detailed documentation to package level.
Code blocks don't work well in godoc for variables because they render the same
as the "var Name = Value" blocks themselves, resulting in a confusing look.

Also make the instructions in PACKAGING.md even briefer, now that the full
instruction is available as a continuous prose in the godoc for pkg/buildinfo.