Remove support for braced lists in lvalue positions.
This was necessary for the legacy assignment syntax when the RHS evaluates to
multiple values, so the only choice is to cram multiple variable names into a
braced list, as in "{a,b}=(cmd) echo $a $b". This is no longer a problem with
the newer assignment commands, like "with a b = (cmd) { echo $a $b }".

This is a fairly obscure feature, and only necessary with the legacy assignment
syntax. Code like `var {a,b} = (cmd)` would now be broken, but a search for `}=`
on GitHub has only found files that also use the legacy assignment syntax and is
broken by its removal anyway. It seems safe enough to remove directly without
deprecation or documenting in the release notes.