pkg/shell: Simplify the logic to find run directory.
Elvish uses a subdirectory under XDG_RUNTIME_DIR if the environment variable is
non-empty, but falls back to a subdirectory of tmpdir it it is unset or empty.

The tmpdir-based one used to be the only one. When support for XDG_RUNTIME_DIR
was added in 0.14.0, to ensure a smooth transition, Elvish would still use the
tmpdir-based one if it exists.

Enough time has passed since 0.14.0, and we can now simplify the run directory
to be solely determined by XDG_RUNTIME_DIR.

Also convert the test to a transcript test.