Introduce pkg/etk, a new TUI framework.
The framework has the following interesting features:

- It has a first-class Elvish binding with the same expressive power as the Go
  version of the framework.

- It has an immediate-mode API, meaning that a component is represented by a
  function that gets invoked every time the state changes.

For more on the motivation and tradeoffs of the design, see the slidedeck in
website/slides/draft-etk.md.

This package will eventually replace pkg/cli/tk as Elvish's internal TUI
framework, and Elvish's TUI will be rewritten to use it.

This commit also incorporates the following changes that overlapped with the code change necessary for implementing etk:

- pkg/ui: De-export TextFromSegment. The styled builtin used to require it, but no longer does now that Text.Scan is implemented.