Add Elvish binding for Go interfaces.
A Go interface can now be implemented by an Elvish map of functions.
For example, interface{ Foo(a int) string } can be implemented by an Elvish map [&foo-impl=...].
One way to think of this is that Go interfaces can now behave like TypeScript interfaces in Elvish-land.

This is unfortunately not very seamless:
Go doesn't support creating new interface implementations dynamically with reflection,
so we have to create a "struct of func" implementation for each interface,
where each field is a function that implements a method,
and convert the Elvish map into that struct
(using the existing structmap mechanism).

The name of a field in a "struct of func" has to be different from the method it implements,
so adopt the convention of an "Impl" suffix.
Since we want the Elvish map to not have an "-impl" suffix,
also implement the ability for a structmap to override the Elvish field name using a field tag.

This is particularly useful for Etk bindings,
and this commit contains a simple elvts test for the combobox component,
whose gen-list state var is expected to return an interface.